cse-498-team-msu-rissar/Local Website/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
eventlet==0.35.1
gunicorn==21.2.0
Flask-Failsafe==0.2
requests==2.31.0
numpy==1.26.2
matplotlib==3.8.2
bokeh==3.3.2
websocket-client==1.7.0
docker==6.1.3
paramiko==3.4.0
